requirements.txt
python-telegram-bot==13.7
telegram
aiohttp
lxml_html_clean
Flask
gunicorn
werkzeug
aioshutil
aiofiles
anitopy
schedule
apscheduler
dnspython
motor
feedparser
psutil
pycountry
pymongo
pyrofork==2.3.45
pyromod==1.5
python-dotenv
tgcrypto
html-telegraph-poster
uvloop
torrentp
ffmpeg
asyncio
pytz
requests
bs4
